requirements.txt
python-telegram-bot
google-genai
datetime